% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scheduleDetail.R
\name{scheduleDetailUI}
\alias{scheduleDetailUI}
\title{生产订单排产信息明细表}
\usage{
scheduleDetailUI(
  tabTitle = "生产订单排产明细表",
  colTitles = c("操作区域", "报表区域"),
  widthRates = c(5, 7),
  func_left = scheduleDetailFunc_left,
  func_right = scheduleDetailFunc_right
)
}
\arguments{
\item{tabTitle}{标题}

\item{com_left}{左}

\item{com_right}{右}
}
\value{
返回值
}
\description{
生产订单排产信息明细表
}
\examples{
scheduleDetailUI()
}
